buildly---project-Based-learning-platform-master/backendPBL/projectBPL/requirements.txt
# requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-decouple # لإدارة الإعدادات
django-cors-headers  # للتواصل مع Frontend